+++ arm-sysregs/examples/log_all.rs
@@ -711,6 +711,11 @@ fn entry(_: u64, _: u64, _: u64, _: u64) -> ! {
     );
     #[cfg(all(any(test, feature = "fakes", target_arch = "aarch64"), feature = "el1"))]
     info!(
+        "id_aa64isar3_el1 = {:?}",
+        arm_sysregs::read_id_aa64isar3_el1()
+    );
+    #[cfg(all(any(test, feature = "fakes", target_arch = "aarch64"), feature = "el1"))]
+    info!(
         "id_aa64mmfr0_el1 = {:?}",
         arm_sysregs::read_id_aa64mmfr0_el1()
     );